backend/requirements.txt
fastapi
uvicorn
psycopg2
requests
python-dotenv
flask
gunicorn
mysql-connector-python
flask-cors


